Making stuff build with MS Visual Studio .NET; probably has some problems still, but should be closer
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2084 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/RenderOther/FontHandler.cpp
@@ -276,7 +276,7 @@ void FontSpecifier::OGL_Reset(bool IsStarting)
 	
 	int GlyphHeight = ascent_p + descent_p;
 	
-	int EstDim = int(sqrt(TotalWidth*GlyphHeight) + 0.5);
+	int EstDim = int(sqrt(static_cast<float>(TotalWidth*GlyphHeight)) + 0.5);
 	TxtrWidth = MAX(128, NextPowerOfTwo(EstDim));
 	
 	// Find the character starting points and counts
